refactor http client (part 2)
