+++ src/Projections/ProjectionsManager.php
@@ -13,36 +13,48 @@ declare(strict_types=1);
 
 namespace Prooph\EventStoreHttpClient\Projections;
 
-use Prooph\EventStore\EndPoint;
+use Http\Message\RequestFactory;
+use Prooph\EventStore\Exception\EventStoreConnectionException;
 use Prooph\EventStore\Exception\InvalidArgumentException;
 use Prooph\EventStore\Projections\ProjectionDetails;
 use Prooph\EventStore\Projections\ProjectionsManager as SyncProjectionsManager;
-use Prooph\EventStore\Transport\Http\EndpointExtensions;
+use Prooph\EventStore\Transport\Http\HttpStatusCode;
 use Prooph\EventStore\UserCredentials;
+use Prooph\EventStore\Util\Json;
+use Prooph\EventStoreHttpClient\ConnectionSettings;
+use Prooph\EventStoreHttpClient\Exception\ProjectionCommandConflictException;
+use Prooph\EventStoreHttpClient\Exception\ProjectionCommandFailedException;
 use Prooph\EventStoreHttpClient\Http\HttpClient;
+use Psr\Http\Client\ClientInterface;
+use Psr\Http\Message\ResponseInterface;
+use Throwable;
 
 class ProjectionsManager implements SyncProjectionsManager
 {
-    /** @var ProjectionsClient */
-    private $client;
-    /** @var EndPoint */
-    private $httpEndPoint;
-    /** @var string */
-    private $httpSchema;
-    /** @var UserCredentials|null */
-    private $defaultUserCredentials;
+    /** @var HttpClient */
+    private $httpClient;
+    /** @var ConnectionSettings */
+    private $settings;
 
     /** @internal */
     public function __construct(
-        HttpClient $client,
-        EndPoint $endPoint,
-        string $schema = EndpointExtensions::HTTP_SCHEMA,
-        ?UserCredentials $defaultUserCredentials = null
+        ClientInterface $client,
+        RequestFactory $requestFactory,
+        ConnectionSettings $settings
     ) {
-        $this->client = new ProjectionsClient($client);
-        $this->httpEndPoint = $endPoint;
-        $this->httpSchema = $schema;
-        $this->defaultUserCredentials = $defaultUserCredentials;
+        $this->settings = $settings;
+
+        $this->httpClient = new HttpClient(
+            $client,
+            $requestFactory,
+            $settings,
+            \sprintf(
+                '%s://%s:%s',
+                $settings->schema(),
+                $settings->endPoint()->host(),
+                $settings->endPoint()->port()
+            )
+        );
     }
 
     /**
@@ -54,11 +66,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        $this->client->enable(
-            $this->httpEndPoint,
-            $name,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+        $this->sendPost(
+            \sprintf(
+                '/projection/%s/command/enable',
+                \urlencode($name)
+            ),
+            '',
+            $userCredentials,
+            HttpStatusCode::OK
         );
     }
 
@@ -71,11 +86,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        $this->client->disable(
-            $this->httpEndPoint,
-            $name,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+        $this->sendPost(
+            \sprintf(
+                '/projection/%s/command/disable',
+                \urlencode($name)
+            ),
+            '',
+            $userCredentials,
+            HttpStatusCode::OK
         );
     }
 
@@ -88,11 +106,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        $this->client->abort(
-            $this->httpEndPoint,
-            $name,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+        $this->sendPost(
+            \sprintf(
+                '/projection/%s/command/abort',
+                \urlencode($name)
+            ),
+            '',
+            $userCredentials,
+            HttpStatusCode::OK
         );
     }
 
@@ -101,17 +122,21 @@ class ProjectionsManager implements SyncProjectionsManager
      */
     public function createOneTime(
         string $query,
+        string $type = 'JS',
         ?UserCredentials $userCredentials = null
     ): void {
         if ('' === $query) {
             throw new InvalidArgumentException('Query is required');
         }
 
-        $this->client->createOneTime(
-            $this->httpEndPoint,
+        $this->sendPost(
+            \sprintf(
+                '/projections/onetime?type=%s',
+                $type
+            ),
             $query,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+            $userCredentials,
+            HttpStatusCode::CREATED
         );
     }
 
@@ -121,6 +146,7 @@ class ProjectionsManager implements SyncProjectionsManager
     public function createTransient(
         string $name,
         string $query,
+        string $type = 'JS',
         ?UserCredentials $userCredentials = null
     ): void {
         if ('' === $name) {
@@ -131,12 +157,15 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Query is required');
         }
 
-        $this->client->createTransient(
-            $this->httpEndPoint,
-            $name,
+        $this->sendPost(
+            \sprintf(
+                '/projections/transient?name=%s&type=%s',
+                \urlencode($name),
+                $type
+            ),
             $query,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+            $userCredentials,
+            HttpStatusCode::CREATED
         );
     }
 
@@ -147,6 +176,7 @@ class ProjectionsManager implements SyncProjectionsManager
         string $name,
         string $query,
         bool $trackEmittedStreams = false,
+        string $type = 'JS',
         ?UserCredentials $userCredentials = null
     ): void {
         if ('' === $name) {
@@ -157,13 +187,16 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Query is required');
         }
 
-        $this->client->createContinuous(
-            $this->httpEndPoint,
-            $name,
+        $this->sendPost(
+            \sprintf(
+                '/projections/continuous?name=%s&type=%s&emit=1&trackemittedstreams=%d',
+                \urlencode($name),
+                $type,
+                (int) $trackEmittedStreams
+            ),
             $query,
-            $trackEmittedStreams,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+            $userCredentials,
+            HttpStatusCode::CREATED
         );
     }
 
@@ -174,11 +207,25 @@ class ProjectionsManager implements SyncProjectionsManager
      */
     public function listAll(?UserCredentials $userCredentials = null): array
     {
-        return $this->client->listAll(
-            $this->httpEndPoint,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+        $response = $this->sendGet(
+            '/projections/any',
+            $userCredentials,
+            HttpStatusCode::OK
         );
+
+        $data = Json::decode($response->getBody()->getContents());
+
+        $projectionDetails = [];
+
+        if (null === $data['projections']) {
+            return $projectionDetails;
+        }
+
+        foreach ($data['projections'] as $entry) {
+            $projectionDetails[] = $this->buildProjectionDetails($entry);
+        }
+
+        return $projectionDetails;
     }
 
     /**
@@ -188,11 +235,25 @@ class ProjectionsManager implements SyncProjectionsManager
      */
     public function listOneTime(?UserCredentials $userCredentials = null): array
     {
-        return $this->client->listOneTime(
-            $this->httpEndPoint,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+        $response = $this->sendGet(
+            '/projections/onetime',
+            $userCredentials,
+            HttpStatusCode::OK
         );
+
+        $data = Json::decode($response->getBody()->getContents());
+
+        $projectionDetails = [];
+
+        if (null === $data['projections']) {
+            return $projectionDetails;
+        }
+
+        foreach ($data['projections'] as $entry) {
+            $projectionDetails[] = $this->buildProjectionDetails($entry);
+        }
+
+        return $projectionDetails;
     }
 
     /**
@@ -202,11 +263,25 @@ class ProjectionsManager implements SyncProjectionsManager
      */
     public function listContinuous(?UserCredentials $userCredentials = null): array
     {
-        return $this->client->listContinuous(
-            $this->httpEndPoint,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+        $response = $this->sendGet(
+            '/projections/continuous',
+            $userCredentials,
+            HttpStatusCode::OK
         );
+
+        $data = Json::decode($response->getBody()->getContents());
+
+        $projectionDetails = [];
+
+        if (null === $data['projections']) {
+            return $projectionDetails;
+        }
+
+        foreach ($data['projections'] as $entry) {
+            $projectionDetails[] = $this->buildProjectionDetails($entry);
+        }
+
+        return $projectionDetails;
     }
 
     /**
@@ -218,12 +293,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        return $this->client->getStatus(
-            $this->httpEndPoint,
-            $name,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
-        );
+        return $this->sendGet(
+            \sprintf(
+                '/projection/%s',
+                \urlencode($name)
+            ),
+            $userCredentials,
+            HttpStatusCode::OK
+        )->getBody()->getContents();
     }
 
     /**
@@ -235,12 +312,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        return $this->client->getState(
-            $this->httpEndPoint,
-            $name,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
-        );
+        return $this->sendGet(
+            \sprintf(
+                '/projection/%s/state',
+                \urlencode($name)
+            ),
+            $userCredentials,
+            HttpStatusCode::OK
+        )->getBody()->getContents();
     }
 
     /**
@@ -259,13 +338,15 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Partition is required');
         }
 
-        return $this->client->getPartitionState(
-            $this->httpEndPoint,
-            $name,
-            $partition,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
-        );
+        return $this->sendGet(
+            \sprintf(
+                '/projection/%s/state?partition=%s',
+                \urlencode($name),
+                \urlencode($partition)
+            ),
+            $userCredentials,
+            HttpStatusCode::OK
+        )->getBody()->getContents();
     }
 
     /**
@@ -277,12 +358,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        return $this->client->getResult(
-            $this->httpEndPoint,
-            $name,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
-        );
+        return $this->sendGet(
+            \sprintf(
+                '/projection/%s/result',
+                \urlencode($name)
+            ),
+            $userCredentials,
+            HttpStatusCode::OK
+        )->getBody()->getContents();
     }
 
     /**
@@ -301,13 +384,15 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Partition is required');
         }
 
-        return $this->client->getPartitionResult(
-            $this->httpEndPoint,
-            $name,
-            $partition,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
-        );
+        return $this->sendGet(
+            \sprintf(
+                '/projection/%s/result?partition=%s',
+                \urlencode($name),
+                \urlencode($partition)
+            ),
+            $userCredentials,
+            HttpStatusCode::OK
+        )->getBody()->getContents();
     }
 
     /**
@@ -319,12 +404,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        return $this->client->getStatistics(
-            $this->httpEndPoint,
-            $name,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
-        );
+        return $this->sendGet(
+            \sprintf(
+                '/projection/%s/statistics',
+                \urlencode($name)
+            ),
+            $userCredentials,
+            HttpStatusCode::OK
+        )->getBody()->getContents();
     }
 
     public function getQuery(string $name, ?UserCredentials $userCredentials = null): string
@@ -333,12 +420,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        return $this->client->getQuery(
-            $this->httpEndPoint,
-            $name,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
-        );
+        return $this->sendGet(
+            \sprintf(
+                '/projection/%s/query',
+                \urlencode($name)
+            ),
+            $userCredentials,
+            HttpStatusCode::OK
+        )->getBody()->getContents();
     }
 
     public function updateQuery(
@@ -355,13 +444,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Query is required');
         }
 
-        $this->client->updateQuery(
-            $this->httpEndPoint,
-            $name,
+        $this->sendPut(
+            \sprintf(
+                '/projection/%s/query?emit=' . (int) $emitEnabled,
+                \urlencode($name)
+            ),
             $query,
-            $emitEnabled,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+            $userCredentials,
+            HttpStatusCode::OK
         );
     }
 
@@ -374,12 +464,14 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        $this->client->delete(
-            $this->httpEndPoint,
-            $name,
-            $deleteEmittedStreams,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+        $this->sendDelete(
+            \sprintf(
+                '/projection/%s?deleteEmittedStreams=%d',
+                \urlencode($name),
+                (int) $deleteEmittedStreams
+            ),
+            $userCredentials,
+            HttpStatusCode::OK
         );
     }
 
@@ -389,11 +481,163 @@ class ProjectionsManager implements SyncProjectionsManager
             throw new InvalidArgumentException('Name is required');
         }
 
-        $this->client->reset(
-            $this->httpEndPoint,
-            $name,
-            $userCredentials ?? $this->defaultUserCredentials,
-            $this->httpSchema
+        $this->sendPost(
+            \sprintf(
+                '/projection/%s/command/reset',
+                \urlencode($name)
+            ),
+            '',
+            $userCredentials,
+            HttpStatusCode::OK
         );
+    }
+
+    private function buildProjectionDetails(array $entry): ProjectionDetails
+    {
+        return new ProjectionDetails(
+            $entry['coreProcessingTime'],
+            $entry['version'],
+            $entry['epoch'],
+            $entry['effectiveName'],
+            $entry['writesInProgress'],
+            $entry['readsInProgress'],
+            $entry['partitionsCached'],
+            $entry['status'],
+            $entry['stateReason'],
+            $entry['name'],
+            $entry['mode'],
+            $entry['position'],
+            $entry['progress'],
+            $entry['lastCheckpoint'],
+            $entry['eventsProcessedAfterRestart'],
+            $entry['statusUrl'],
+            $entry['stateUrl'],
+            $entry['resultUrl'],
+            $entry['queryUrl'],
+            $entry['enableCommandUrl'],
+            $entry['disableCommandUrl'],
+            $entry['checkpointStatus'],
+            $entry['bufferedEvents'],
+            $entry['writePendingEventsBeforeCheckpoint'],
+            $entry['writePendingEventsAfterCheckpoint']
+        );
+    }
+
+    private function sendGet(
+        string $uri,
+        ?UserCredentials $userCredentials,
+        int $expectedCode
+    ): ResponseInterface {
+        $response = $this->httpClient->get(
+            $uri,
+            [],
+            $userCredentials,
+            static function (Throwable $e) {
+                throw new EventStoreConnectionException($e->getMessage());
+            }
+        );
+
+        if ($response->getStatusCode() !== $expectedCode) {
+            throw new ProjectionCommandFailedException(
+                $response->getStatusCode(),
+                \sprintf(
+                    'Server returned %d (%s) for GET on %s',
+                    $response->getStatusCode(),
+                    $response->getReasonPhrase(),
+                    $uri
+                )
+            );
+        }
+
+        return $response;
+    }
+
+    private function sendDelete(
+        string $uri,
+        ?UserCredentials $userCredentials,
+        int $expectedCode
+    ): void {
+        $response = $this->httpClient->delete(
+            $uri,
+            [],
+            $userCredentials,
+            static function (Throwable $e) {
+                throw new EventStoreConnectionException($e->getMessage());
+            }
+        );
+
+        if ($response->getStatusCode() !== $expectedCode) {
+            throw new ProjectionCommandFailedException(
+                $response->getStatusCode(),
+                \sprintf(
+                    'Server returned %d (%s) for DELETE on %s',
+                    $response->getStatusCode(),
+                    $response->getReasonPhrase(),
+                    $uri
+                )
+            );
+        }
+    }
+
+    private function sendPut(
+        string $uri,
+        string $content,
+        ?UserCredentials $userCredentials,
+        int $expectedCode
+    ): void {
+        $response = $this->httpClient->put(
+            $uri,
+            ['Content-Type' => 'application/json'],
+            $content,
+            $userCredentials,
+            static function (Throwable $e) {
+                throw new EventStoreConnectionException($e->getMessage());
+            }
+        );
+
+        if ($response->getStatusCode() !== $expectedCode) {
+            throw new ProjectionCommandFailedException(
+                $response->getStatusCode(),
+                \sprintf(
+                    'Server returned %d (%s) for PUT on %s',
+                    $response->getStatusCode(),
+                    $response->getReasonPhrase(),
+                    $uri
+                )
+            );
+        }
+    }
+
+    private function sendPost(
+        string $uri,
+        string $content,
+        ?UserCredentials $userCredentials,
+        int $expectedCode
+    ): void {
+        $response = $this->httpClient->post(
+            $uri,
+            ['Content-Type' => 'application/json'],
+            $content,
+            $userCredentials,
+            static function (Throwable $e) {
+                throw new EventStoreConnectionException($e->getMessage());
+            }
+        );
+
+        if ($response->getStatusCode() === HttpStatusCode::CONFLICT) {
+            throw new ProjectionCommandConflictException($response->getStatusCode(), $response->getReasonPhrase());
+        }
+
+        if ($response->getStatusCode() !== $expectedCode) {
+            throw new ProjectionCommandFailedException(
+                $response->getStatusCode(),
+                \sprintf(
+                    'Server returned %d (%s) for POST on %s',
+                    $response->getStatusCode(),
+                    $response->getReasonPhrase(),
+                    $uri
+                )
+            );
+        }
     }
 }
