+++ src/Projections/QueryManager.php
@@ -13,11 +13,11 @@ declare(strict_types=1);
 
 namespace Prooph\EventStoreHttpClient\Projections;
 
-use Prooph\EventStore\EndPoint;
+use Http\Message\RequestFactory;
 use Prooph\EventStore\Projections\QueryManager as SyncQueryManager;
-use Prooph\EventStore\Transport\Http\EndpointExtensions;
 use Prooph\EventStore\UserCredentials;
-use Prooph\EventStoreHttpClient\Http\HttpClient;
+use Prooph\EventStoreHttpClient\ConnectionSettings;
+use Psr\Http\Client\ClientInterface;
 
 /**
  * API for executing queries in the Event Store through PHP code.
@@ -30,18 +30,16 @@ class QueryManager implements SyncQueryManager
     /** @var ProjectionsManager */
     private $projectionsManager;
 
-    /** @internal  */
+    /** @internal */
     public function __construct(
-        HttpClient $client,
-        EndPoint $endPoint,
-        string $schema = EndpointExtensions::HTTP_SCHEMA,
-        ?UserCredentials $defaultUserCredentials = null
+        ClientInterface $client,
+        RequestFactory $requestFactory,
+        ConnectionSettings $settings
     ) {
         $this->projectionsManager = new ProjectionsManager(
             $client,
-            $endPoint,
-            $schema,
-            $defaultUserCredentials
+            $requestFactory,
+            $settings
         );
     }
 
@@ -56,6 +54,7 @@ class QueryManager implements SyncQueryManager
      * @param string $query The source code for the query
      * @param int $initialPollingDelay Initial time to wait between polling for projection status
      * @param int $maximumPollingDelay Maximum time to wait between polling for projection status
+     * @param string $type The type to use, defaults to JS
      * @param UserCredentials|null $userCredentials Credentials for a user with permission to create a query
      *
      * @return string
@@ -65,11 +64,13 @@ class QueryManager implements SyncQueryManager
         string $query,
         int $initialPollingDelay,
         int $maximumPollingDelay,
+        string $type = 'JS',
         ?UserCredentials $userCredentials = null
     ): string {
         $this->projectionsManager->createTransient(
             $name,
             $query,
+            $type,
             $userCredentials
         );
